Add MDIT for the installation guide. (#35)
This was made as lynx, our current solution, doesn't have the correct URL, and even if it did, lynx doesn't parse javascript. This fixes the issue by using MDIT, one of my tools. MDIT is a seperate (at the time of writing) a 135 line long bash script which at least for now will work better than lynx for this. We should keep lynx however, as a web browser in any case is useful.